feat: move yaml functions to utils.py
